requirements.txt
matplotlib==3.5.1
numpy==1.22.0
pandas==1.3.5
Pillow==9.4.0
torch==1.11.0+cu113
torchsummary==1.5.1
torchvision==0.12.0+cu113
